Add second enemy

+++ Assets/Scripts/ScriptableObjects/ExplosionTowerScriptableObject.cs
@@ -6,13 +6,13 @@ namespace ScriptableObjects
     public class ExplosionTowerScriptableObject : BaseTowerScriptableObject
     {
         [Header("Level 1")]
-        public float explosionDamage1;
+        public int explosionDamage1;
         public float explosionRadius1;
         [Header("Level 2")]
-        public float explosionDamage2;
+        public int explosionDamage2;
         public float explosionRadius2;
         [Header("Level 3")]
-        public float explosionDamage3;
+        public int explosionDamage3;
         public float explosionRadius3;
     }
 }
